src/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.2.1
django-bootstrap4==23.1
django-crispy-forms==2.0
djangorestframework==3.14.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
